Create raw data logger.
